+++ activerecord/lib/active_record/connection_adapters/sqlite3_adapter.rb
@@ -694,7 +694,7 @@ module ActiveRecord
                 column_options[:stored] = column.virtual_stored?
                 column_options[:type] = column.type
               elsif column.has_default?
-                default = column.cast_type.deserialize(column.default)
+                default = column.cast_type.deserialize(column.default_before_type_cast)
                 default = -> { column.default_function } if default.nil?
 
                 unless column.auto_increment?